+++ src/services/CompareUtils.jl
@@ -1,7 +1,9 @@
+##==============================================================================
+## (==)
+##==============================================================================
+import Base.==
 ## @generated compare
 # Reference https://github.com/JuliaLang/julia/issues/4648
-
-import Base.==
 
 #=
 For now abstract `InferenceVariable`s are considered equal if they are the same type, dims, and manifolds (abels are deprecated)
@@ -14,7 +16,6 @@ implement compare if needed.
 
 ==(a::FunctorInferenceType, b::FunctorInferenceType) = typeof(a) == typeof(b)
 
-# TestCCW1{TestFunctorInferenceType1}
 
 # Generate compares automatically for all in this union
 const GeneratedCompareUnion = Union{MeanMaxPPE, VariableNodeData, DFGNodeParams,
@@ -27,7 +28,9 @@ const GeneratedCompareUnion = Union{MeanMaxPPE, VariableNodeData, DFGNodeParams,
 end
 
 
-##
+##==============================================================================
+## Compare
+##==============================================================================
 
 function compareField(Allc, Bllc, syms)::Bool
     (!isdefined(Allc, syms) && !isdefined(Bllc, syms)) && return true
